Test by comparing with expected output
