+++ vm_server/test/execute_macro/code/execute.py
@@ -19,13 +19,15 @@ def execute_macro():
     xl_file = win32com.client.Dispatch("Excel.Application")
     xl_run = xl_file.Workbooks.Open(os.path.abspath(path_to_file),
                                     ReadOnly=1)
-    xl_run.Application.Run("excelsheet.xlsm!main.simpleMain") #execute macro
+    xl_run.Application.Run("excelsheet.xlsm!Module1.add_numbers_in_column") #execute macro
     xl_run.Save()
     xl_run.Close()
     xl_file.Quit()
     del xl_file
     shutil.move(path_to_file, current_path +
                 "\\action\\output\\excelsheet.xlsm")
+    shutil.move(current_path + "\\action\\data\\output.txt", current_path +
+                "\\action\\output\\output.txt")
     print("Action successfully executed")
 
 if __name__ == "__main__":
