+++ vm_server/test/execute_macro/expected_output/expected_output.txt
@@ -1,0 +1,1 @@
+21091
